pyproject.toml
[project]
name = "askbro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.0",
    "environs>=14.3.0",
    "injector>=0.22.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.2",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.44",
    "undefined>=0.0.8",
    "valkey[libvalkey]>=6.1.1",
    "xai-sdk>=1.3.1",
    "greenlet>=3.2.4",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.0",
]
